LayoutTests/fast/mediastream/RTCPeerConnection-ice-expected.txt

Tests the RTCPeerConnection Ice functionality.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pc = new webkitRTCPeerConnection(null, null); did not throw exception.
PASS iceConnectionState is completed
PASS onIceCandidate was called.
PASS event.candidate.candidate is "nano nano"
PASS iceConnectionState is closed.
PASS successfullyParsed is true

TEST COMPLETE
